convertisseur/requirements.txt
comtypes
pdfplumber
flask
markdownify
llama_index==0.5.6
torch
datasets
python-pptx
transformers
python-docx
evaluate
accelerate>==0.26.0

